requirements_testing.txt
coverage
coveralls
codecov
git+https://github.com/pytest-dev/pytest/#egg=pytest
pytest-cov
pytest-rerunfailures
